requirements.txt
africastalking==1.2.8
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-rest-framework==0.1.0
django-unfold==0.38.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.0
packaging==24.1
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
schema==0.7.7
service-identity==24.1.0
setuptools==74.1.2
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
zope.interface==7.0.3
